+++ samplesheets/vueapp/src/App.vue
@@ -168,6 +168,12 @@
         ref="columnToggleModalRef">
     </column-toggle-modal>
 
+    <!-- Study/assay table detail modal -->
+    <table-detail-modal
+        :app="getApp()"
+        ref="tableDetailModal">
+    </table-detail-modal>
+
     <!-- Editing: Column configuration modal -->
     <column-config-modal
         v-if="editMode && sodarContext.perms.edit_sheet"
@@ -211,6 +217,7 @@ import IrodsDirModal from './components/modals/IrodsDirModal.vue'
 import StudyShortcutModal from './components/modals/StudyShortcutModal.vue'
 import ColumnToggleModal from './components/modals/ColumnToggleModal.vue'
 import ColumnConfigModal from './components/modals/ColumnConfigModal.vue'
+import TableDetailModal from './components/modals/TableDetailModal.vue'
 import EditorHelpModal from './components/modals/EditorHelpModal.vue'
 import WinExportModal from './components/modals/WinExportModal.vue'
 import OntologyEditModal from './components/modals/OntologyEditModal.vue'
@@ -295,6 +302,7 @@ export default {
     StudyShortcutModal,
     ColumnToggleModal,
     ColumnConfigModal,
+    TableDetailModal,
     EditorHelpModal,
     WinExportModal,
     OntologyEditModal,
@@ -343,7 +351,8 @@ export default {
         projectUuid: this.projectUuid,
         sodarContext: this.sodarContext,
         studyUuid: this.currentStudyUuid,
-        showNotificationCb: this.showNotification
+        showNotificationCb: this.showNotification,
+        tableDetailModal: this.$refs.tableDetailModal
       }
     },